Update noFollowing message in the feed to handle Twitter connect as well.
