+++ apps/feed/client/index.coffee
@@ -2,8 +2,7 @@
 # React requires
 qs = require 'qs'
 { mountWithApolloProvider } = require '../../../react/apollo/index.js'
-{ default: EmptyConnectTwitter } = require '../../../react/pages/feed/components/EmptyConnectTwitter/index.js'
-{ default: NoFollowingMessage } = require '../../../react/pages/feed/components/NoFollowingMessage/index.js'
+{ default: NoFollowingMessage } = require '../../../react/components/Feed/components/NoFollowingMessage/index.js'
 
 # Legacy requires
 { FEED_TYPE } = require('sharify').data
@@ -11,9 +10,6 @@ qs = require 'qs'
 { default: initializeFeed } = require './feed.js'
 
 module.exports = ->
-  if ($emptyComponent = $('.js-empty-feed')).length
-    return mountWithApolloProvider(EmptyConnectTwitter, null, $emptyComponent)
-
   if ($noFollowingComponent = $('.js-no-following')).length
     mountWithApolloProvider(NoFollowingMessage, { }, $noFollowingComponent)
