+++ react/components/Feed/components/NoFollowingMessage/queries/followerCount.js
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import FollowerCountCheck from 'react/pages/feed/components/NoFollowingMessage/fragments/followerCount';
+import FollowerCountCheck from 'react/components/Feed/components/NoFollowingMessage/fragments/followerCount';
 
 export default gql`
   query FollowerCountCheckQuery {
